Adding search activity (WIP)

+++ app/src/main/java/com/ayova/moviseries/models/LastEpisodeToAir.kt
@@ -10,6 +10,6 @@ data class LastEpisodeToAir(
     val season_number: Int,
     val show_id: Int,
     val still_path: String,
-    val vote_average: Int,
+    val vote_average: Double,
     val vote_count: Int
 )
